pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Lutf", email = "oolucaslutf@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=67.6.0",
    "python-magic>=0.4.27",
    "Pillow>=9.4.0",
    "drf-spectacular>=0.26.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
